Add files via upload
The first batch of robots is ready, now they must be tested. All robots are unique, so the speed of movement is different for each robot. 
We have to determine the minimum, maximum, and average the robots' speed.
Our server gets an array of testResults robot speeds, passes it through the getSpeedStatistic function and returns the statistics in an array.
The first element of the resulting array is the slowest speed
the second element is the highest speed
the third is average speed rounded down