Add files via upload
Robots tested. Slow robots sent for upgrade. Mate Post wants to buy a dozen robots from us to deliver goods around the city. 
In order to do that. They need the robots that will carry the most cargo in a day.
Let's write a compareRobots function that gets two arrays. 
The first one is firstRobotResults array is the weights of cargo moved during the day by the first robot, the secondRobotResults array is the second one.
Check which of the robots has carried more weight, and return the string with a recommendation, which of the robots is worth buying