Add files via upload
Let's check if all our numbers are "neat" and all the digits of the number are in the right order. The order is correct when all digits of a number are in // non-decreasing sequence.  in a non-decreasing sequence.

Create an isTidy function that takes number as a parameter and // should output the result true if should print the result true if the digits are in non-decreasing order, or false if not.

Note: Input values are always positive.